+++ src/components/admin/menu/MenuContainer.js
@@ -3,6 +3,10 @@ import MenuForm from "./MenuForm";
 import {connect} from "react-redux";
 import {deleteMenuById, fetchMenu, getMenuById, submitMenu} from "../../../services/MenuService";
 import {editMenuForm, fetchMenuSuccess, resetMenuForm} from "./MenuAction";
+import Swal from 'sweetalert2'
+import withReactContent from 'sweetalert2-react-content'
+
+const MySwal = withReactContent(Swal);
 
 class MenuContainer extends React.Component {
 
@@ -77,8 +81,20 @@ class MenuContainer extends React.Component {
     };
 
     handleDeleteData = async (id) => {
-        await deleteMenuById(id)
-            .then(this.fetchDataMenu);
+        MySwal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then( async (result) => {
+            if (result.value){
+                await deleteMenuById(id)
+                    .then(this.fetchDataMenu);
+            }
+        })
     }
 }
